Add ability to close session
