+++ README.md
@@ -41,13 +41,18 @@ password = "Your password"
 async def main():
     # Create an account.
     account = Account()
-    # Connect to the API and load robots
-    await account.connect(username=username, password=password, load_robots=True)
 
-    # Print robots associated with account.
-    print("Robots:")
-    for robot in account.robots:
-        print(robot)
+    try:
+        # Connect to the API and load robots.
+        await account.connect(username=username, password=password, load_robots=True)
+
+        # Print robots associated with account.
+        print("Robots:")
+        for robot in account.robots:
+            print(robot)
+    finally:
+        # Disconnect from the API.
+        await account.disconnect()
 
 
 if __name__ == "__main__":
